feat(skills): move module runtime analysis into plaited-context
Add deterministic module-pattern and module-flow scripts under plaited-context,
including optional SQLite recording through existing findings tables.

Update context assembly to recommend module analysis commands and TypeScript-LSP
probes for module actor review tasks while preserving source authority ordering.

Convert mss-module-review scripts into compatibility wrappers backed by the new
plaited-context scripts, and update skill docs/tests to reflect the compatibility role.
